fix: updated hashing algorithm to probe all slots

+++ lib/hash/srcs/pow_ll.c
@@ -6,13 +6,13 @@
 /*   By: lravier <lravier@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/04/29 18:08:38 by lravier       #+#    #+#                 */
-/*   Updated: 2020/04/29 18:12:51 by lravier       ########   odam.nl         */
+/*   Updated: 2020/06/12 17:21:19 by lravier       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/hashing.h"
 
-long long unsigned	ft_pow_ll(int base, int pow)
+long long unsigned	ft_pow_ll(unsigned long long base, unsigned long long pow)
 {
 	if (pow == 0)
 		return (1);
